refactor: extract PouchDbContext types to separate file and fix test console.error suppression
