+++ packages/client/src/hooks/use_database_health.test.tsx
@@ -4,7 +4,7 @@ import { act, renderHook } from '@testing-library/react';
 import React from 'react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { PouchDbContext, type PouchDbContextType } from '../pouch_db';
+import { PouchDbContext, type PouchDbContextType } from '../pouch_db_types';
 import '../test-polyfill';
 import { createTestPouchDb, destroyTestPouchDb } from '../test-setup';
 import { useDatabaseHealth } from './use_database_health';
